backend: organise generateStarSystem
